- Added: Support for abilities to be loaded onto NPC templates, and for Pets to use Intercept only (for now) - Bugfix: Fixed various gain experience calls which shouldn't be modified by the XP modifier. - Bugfix: By default guild rank 9 can speak in guild chat. - Bugfix: You can now delve Speed of the Realm. - Changed: GMs will no longer be created in starter guilds. - Added: Server property for freelevel_days, set to -1 if you want to deactivate freelevels. - Bugfix: You can no longer jump to someone who is inside a house.

+++ GameServerScripts/playercommands/level.cs
@@ -40,7 +40,7 @@ namespace DOL.GS.Scripts
 							long newXP;
 							//calculate xp to level 20 from current level..
 							newXP = GamePlayer.XPLevel[20] - client.Player.Experience;
-							newXP = (long)((double)newXP / ServerProperties.Properties.XP_RATE);
+							newXP = (long)(newXP / ServerProperties.Properties.XP_RATE);
 							if (newXP < 0) newXP = 0;
 							client.Player.GainExperience(newXP);
 							client.Player.PlayerCharacter.UsedLevelCommand = true;
